refactor(mutation): own transaction, demote edit-engine
Move persist-undo → write → restore transaction from edit-engine
into Mutation as private helpers commitSingle/commitBatch with
per-caller undoUnavailableMessage and restore policy. The boolean
restoreUnwrittenUndos no longer crosses the Mutation/engine seam;
each caller encodes its policy explicitly.

Demote edit-engine to private src/mutation/engine.ts (applyOne,
runFileEdits, helpers) marked @internal. Keep src/edit-engine.ts
as deprecated shim re-exporting from mutation/engine for existing
tests. Delete src/edit-pipeline.ts shim (export * from mutation).

Stop leaking engine helpers via Mutation re-exports for new code;
deprecated re-exports kept for existing coverage tests to stay
green.
