--- src/edit-pipeline.ts
@@ -1,6 +1,0 @@
-/**
- * @deprecated — private to Mutation seam. Use `from "./mutation.js"` instead.
- * Thin shim re-exporting the pipeline now owned by Mutation.
- * @module dsh-better-edit/edit-pipeline
- */
-export * from "./mutation.js";
